+++ fake_schema_generator/functions/dataclass_to_interface.py
@@ -5,7 +5,16 @@ from typing import Optional
 from typing import get_type_hints
 
 
-def dataclass_to_interface(cls: dataclass) -> dataclass:
+def dataclass_to_interface(cls: type[dataclass]) -> type[dataclass]:
+    """
+    Convert a dataclass to an interface by making all fields optional and defaulting to None
+
+    Args:
+        cls: type[dataclass]: A dataclass to convert to an interface
+
+    Returns:
+        type[dataclass]: A new dataclass type named f"I{cls}" with all fields made optional and defaulting to None
+    """
     new_attrs = {"__annotations__": {}}
     type_hints = get_type_hints(cls)
     for f in fields(cls):