+++ fake_schema_generator/functions/extract_annotations.py
@@ -3,7 +3,16 @@ from typing import Any
 from .resolve_annotated_type import resolve_annotated_type
 
 
-def extract_annotations(cls: type):
+def extract_annotations(cls: type) -> dict[str, Any]:
+    """
+    Extracts type annotations from a type.
+
+    Args:
+        cls (type): The class to extract annotations from
+
+    Returns:
+        dict[str, Any]: A dictionary mapping field names to a dictionary containing the field type and metadata
+    """
     annotations: dict[str, Any] = {}
     for field_name, field_type in cls.__annotations__.items():
         actual_type: tuple[Any, list] = resolve_annotated_type(field_type)