+++ fake_schema_generator/functions/resolve_annotated_type.py
@@ -7,6 +7,15 @@ from typing import get_origin
 
 
 def resolve_annotated_type(cls: type) -> Optional[tuple[Any, list]]:
+    """
+    Unwraps nested `Annotated`s and `TypeAliasType`s to get the base type and metadata.
+
+    Args:
+        cls (type): The class to resolve annotated type from
+
+    Returns:
+        Optional[tuple[Any, list]]: A tuple containing the base type and metadata if the class is annotated, otherwise None
+    """
     origin: ParamSpec = get_origin(cls)
     if origin is Annotated:
         args: tuple[Any, ...] = get_args(cls)