+++ fake_schema_generator/operators/noop_operator.py
@@ -1,2 +1,9 @@
-def noop(*args, **kwargs):
+def noop(*_args, **_kwargs):
+    """
+    An operator that takes all arguments and keyword arguments and does nothing.
+
+    Args:
+        *_args: Any positional arguments.
+        **_kwargs: Any keyword arguments.
+    """
     pass