+++ CHANGELOG.md
@@ -12,7 +12,18 @@ Current development version: [v0.1.0](#v010)
 
 ## Versions
 
+* [v0.1.1](#v011) - 2024-05-03
 * [v0.1.0](#v010) - 2024-05-02
+
+### v0.1.1
+* 🐛 Fixed several places where functions expected `type[dataclass]`, but were hinted with `dataclass` instead
+* ♻️ Removed several instances where entire modules or packages were being imported instead of specific members
+* ♻️ Eliminate the requirement to access "protected" members of `FakeSchemaGenerator`
+  * ✨ Added a `data` method on `FakeSchemaGenerator` to return the generated data
+  * ♻️ Made `FakeSchemaGenerator._register` public as `FakeSchemaGenerator.register`
+* ♻️ Added a `FakeSchemaGenerator.generate` method as a convenience wrapper for `FakeSchemaGenerator.generate_from_dag`
+* ♻️ Modified `SequentialNumberProvider.reset_sequence` to reset all sequences by default instead of one namespace
+* 📝 Documented classes and functions with docstrings, [README.md](README.md) is no longer the only documentation
 
 ### v0.1.0
 * 🐛 `operator.add` and `operator.mul` take two arguments, replaced with `typed_sum` and `typed_product`
@@ -26,7 +37,6 @@ Current development version: [v0.1.0](#v010)
 
 ### Next version
 
-* ♻️ Eliminate the requirement to access "protected" members of `FakeSchemaGenerator`
 * 📦 Package for distribution
 
 ### Future versions
@@ -43,3 +53,4 @@ Current development version: [v0.1.0](#v010)
 * ♻️ Add a way to ensure that each row in a table has at least one corresponding row in another table
   * Specifically, when generating the example schema, there are 100 `Order`s, but many of them have no associated
     `OrderProducts`. However, in a real database, each `Order` would have at least one `OrderProduct`.
+* 🧪 Decide if I care that tests still use protected members of `FakeSchemaGenerator`